1) Update for per residue sasa calculation for the main program, BASH files, and those for all atom per-residue sasa calculation with enerCHARMM.pl. 2) Code for plotting residue-residue distance density distribution. Highlight: main.cpp BASH/calc_all_atom.bash sasa_custom_enerCHARMM.str DIST_HIST

+++ BASH/calc_all_atom.bash
@@ -1,14 +1,24 @@
 #!/bin/bash -l
 module load CHARMM
 module load MMTSB
+
+if [ -f _sasa_all_atom.dat ]
+then
+   rm _sasa_all_atom.dat
+fi
+if [ -f sasa_all_atom.dat ]
+then
+   rm sasa_all_atom.dat
+fi
+
 for i in *.pdb
 do
-  enerCHARMM.pl -out sasa -cmd log.cmd -log $i.log $i
-  name=`echo $i | awk -F "." '{print $1}'`
-  value=`grep 'SURFAC: TOTAL =' $i.log | awk '{print $4}'`
+  name=`echo $i | awk -F '.' '{print $1}'`
+  enerCHARMM.pl -out sasa -cmd log.cmd -custom sasa_custom_enerCHARMM.str -log $name.log $i
+  value=`grep 'SURFAC: TOTAL =' $name.log | awk '{print $4}'`
   printf "$name\t$value\n" >> _sasa_all_atom.dat
-  rm $i.log
+  mv temp_resi.dat sasa_all_"$name"_resi.dat
 done
 
-sort -n -k 1,1 _sasa_all_atom.dat sasa_all_atom.dat
-rm _sasa_all_atom.dat
+sort -n -k 1,1 _sasa_all_atom.dat > sasa_all_atom.dat
+#rm _sasa_all_atom.dat
